scripts/requirements.txt
dashscope>=1.14.0
pyyaml>=6.0
python-slugify>=8.0.0